requirements/dev.txt
-r base.txt

django-extensions==2.2.9
factory-boy==2.12.0
freezegun==0.3.15
ipython==7.14.0
isort==4.3.21
black==19.10b0
# Required in order to run `runserver_plus` (from django-extensions).
Werkzeug==1.0.1
